requirements.txt
pytezos==3.3.4
python-dotenv==0.19.2
pandas==1.4.1
pystan==2.19.1.1
convertdate==2.4.0
prophet==1.0.1
scipy==1.8.0
sympy==1.9
pyarrow==7.0.0
pydantic==1.9.0
httpx==0.22.0
ortools==9.2.9972
more_itertools==8.12.0
pytest==7.0.1
uvicorn==0.17.5
signalrcore==0.9.4
fastapi==0.75.0
httpx==0.22.0
requests==2.27.1
fastapi==0.75.2
python-dotenv==0.20.0
uvicorn==0.17.6